fix: address security and code quality issues from review
- Add HTML escaping for domain names to prevent XSS vulnerabilities
- Move formatTimestamp logic to shared formatters.ts as formatRelativeTime
- Remove duplicate formatTimestamp functions from overview.ts and requests.ts

These are targeted fixes that don't distract from the main PR purpose.
